fix(test): i128 width for standing-total assertion (fab compile-gate)
ledger_total() returns i128 (real Ledger balances are i128); the first
standing-invariant assert compared it against an i64 expression. Surfaced on
the fab rebuild (never compiled locally per no-heavy-builds-on-thinkpad). The
fee-loop SOURCE compiled clean; only this test literal needed the width.

(cherry picked from commit 75424089558468f4ef0dd5d873232669a29c7d70)

+++ turn/tests/faucet_steady_state_solvency.rs
@@ -248,7 +248,8 @@ fn integration_closed_loop_solvent_over_10k_turns() {
     ledger.insert_cell(faucet).unwrap();
 
     let standing_before = ledger_total(&ledger);
-    assert_eq!(standing_before, FAUCET_START + AGENT_BUFFER as i64);
+    // ledger_total() is i128 (real Ledger balances are i128) — match the width.
+    assert_eq!(standing_before, FAUCET_START as i128 + AGENT_BUFFER as i128);
 
     let mut agent_nonce = 0u64;
     let mut faucet_nonce = 0u64;
